feat: round uint64

+++ lib/src/algorand.dart
@@ -268,7 +268,7 @@ class Algorand {
   /// Throws an [AlgorandException] if there is an HTTP error.
   /// Returns the block in the given round number.
   Future<Block> getBlockByRound(
-    int round, {
+    BigInt round, {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
