+++ lib/src/api/block/blocks_algod_api.dart
@@ -21,7 +21,7 @@ class BlocksAlgodApi {
   /// Throws an [AlgorandException] if there is an HTTP error.
   /// Returns the block in the given round number.
   Future<Block> getBlockByRound(
-    int round, {
+    BigInt round, {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
