+++ lib/src/api/block/blocks_indexer_api.dart
@@ -17,7 +17,7 @@ class BlocksIndexerApi {
   /// Throws an [AlgorandException] if there is an HTTP error.
   /// Returns the block in the given round number.
   Future<Block> getBlockByRound(
-    int round, {
+    BigInt round, {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
